backend/requirements.txt
# Flask RAG Pipeline Dependencies
# Using versions with pre-built Windows wheels

fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.7
python-dotenv>=1.0.0

# LangChain Core - using specific compatible versions
langchain==0.2.16
langchain-google-genai<2.0.0
langchain-community==0.2.17
langchain-text-splitters==0.2.4
langchain-chroma>=0.1.0

# Vector Store - using version with Windows wheels
chromadb==0.4.24

# Document Processing
pypdf>=4.0.0
docx2txt>=0.8

# Google Drive
google-api-python-client>=2.100.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
